+++ src/app/watchlists/[watchlistId]/components/WatchlistAnimeTable/WatchlistAnimeTableContainer.tsx
@@ -2,7 +2,7 @@ import { fetchWithType } from '@/lib/api'
 import { proxyRequestHeaders } from '@/lib/headers'
 import { withBaseURL } from '@/lib/url'
 
-import { WatchlistAnimeDataTable } from './WatchlistAnimeDataTable'
+import { WatchlistAnimeTable } from './WatchlistAnimeTable'
 
 import type { GetWatchlistAnimeResponse } from '@/api/watchlists/[watchlistId]/anime/types'
 
@@ -12,7 +12,7 @@ type Props = {
   minWidth?: number
 }
 
-export async function WatchlistAnimeTable({ minWidth, watchlistId, limit }: Props) {
+export async function WatchlistAnimeTableContainer({ minWidth, watchlistId, limit }: Props) {
   const animeResponse = await fetchWithType<GetWatchlistAnimeResponse>(
     withBaseURL(`/api/watchlists/${watchlistId}/anime?limit=${limit}`),
     {
@@ -23,7 +23,7 @@ export async function WatchlistAnimeTable({ minWidth, watchlistId, limit }: Prop
   )
 
   return (
-    <WatchlistAnimeDataTable
+    <WatchlistAnimeTable
       initialData={animeResponse.data?.ok ? animeResponse.data : null}
       watchlistId={watchlistId}
       limit={limit}
