+++ src/app/watchlists/[watchlistId]/components/WatchlistAnimeTable/index.ts
@@ -1,2 +1,2 @@
-export { WatchlistAnimeTable } from './WatchlistAnimeTable'
+export { WatchlistAnimeTableContainer } from './WatchlistAnimeTableContainer'
 export { AnimeTableSkeleton } from '@/components/anime/table/AnimeTableSkeleton'
